Adjust access modifiers and reformat code
Access modifiers were adjusted in the PostgresqlMapperTest for clarity and encapsulation. Code format in DefaultMapper was modified for readability, including reorganizing imports and adding blank lines for separation. Also removed import statement for Optional as it was not being used. Indentation of mapper methods was reduced for better visibility of method hierarchy.