Improve error handling in QueryConfiguration class
The getDefault method in QueryConfiguration class has been improved to throw an exception when called before the DEFAULT is set, instead of returning null. This change makes it clearer to developers that they need to call the setDefault method before calling getDefault, thus preventing potential misuse of the method and enhancing the overall code quality.