Reorganize Adapter and Call interfaces
This commit consists of deletion and re-creation of the `Adapter` and `Call` interfaces, as well as their accompanying classes. The interfaces have been relocated as per the new project structure. Changes include deletion of old interfaces, creation of new interfaces under appropriate packages and modification of corresponding methods and functions.