Move 'Calls.java' and refactor package paths
This commit renames the full path of 'Calls.java', moving it from 'queries/calls' to 'queries/api/call/calls'. This also includes the modification of package paths to maintain consistency and the addition of calls for 'BatchCall' and 'SingletonCall'. These changes improve the organization and visibility of call-related classes within the project structure.